Add alerts manager role and enforce permissions

+++ graylog2-server/src/main/java/org/graylog/events/rest/EventDefinitionsResource.java
@@ -39,6 +39,7 @@ import org.graylog2.search.SearchQuery;
 import org.graylog2.search.SearchQueryField;
 import org.graylog2.search.SearchQueryParser;
 import org.graylog2.shared.rest.resources.RestResource;
+import org.graylog2.shared.security.RestPermissions;
 
 import javax.inject.Inject;
 import javax.validation.Valid;
@@ -92,6 +93,7 @@ public class EventDefinitionsResource extends RestResource implements PluginRest
     public PaginatedResponse<EventDefinitionDto> list(@ApiParam(name = "page") @QueryParam("page") @DefaultValue("1") int page,
                                                       @ApiParam(name = "per_page") @QueryParam("per_page") @DefaultValue("50") int perPage,
                                                       @ApiParam(name = "query") @QueryParam("query") @DefaultValue("") String query) {
+        checkPermission(RestPermissions.EVENT_DEFINITIONS_READ);
         final SearchQuery searchQuery = searchQueryParser.parse(query);
         return PaginatedResponse.create("event_definitions", dbService.getAllPaginated(searchQuery, "title", page, perPage), query);
     }
@@ -100,6 +102,7 @@ public class EventDefinitionsResource extends RestResource implements PluginRest
     @Path("{definitionId}")
     @ApiOperation("Get an event definition")
     public EventDefinitionDto get(@ApiParam(name = "definitionId") @PathParam("definitionId") @NotBlank String definitionId) {
+        checkPermission(RestPermissions.EVENT_DEFINITIONS_READ);
         return dbService.get(definitionId)
                 .orElseThrow(() -> new NotFoundException("Event definition <" + definitionId + "> doesn't exist"));
     }
@@ -109,6 +112,7 @@ public class EventDefinitionsResource extends RestResource implements PluginRest
     @ApiOperation("Create new event definition")
     @AuditEvent(type = EventsAuditEventTypes.EVENT_DEFINITION_CREATE)
     public Response create(EventDefinitionDto dto) {
+        checkPermission(RestPermissions.EVENT_DEFINITIONS_CREATE);
         final ValidationResult result = dto.config().validate();
         if (result.failed()) {
             return Response.status(Response.Status.BAD_REQUEST).entity(result).build();
@@ -122,6 +126,7 @@ public class EventDefinitionsResource extends RestResource implements PluginRest
     @AuditEvent(type = EventsAuditEventTypes.EVENT_DEFINITION_UPDATE)
     public Response update(@ApiParam(name = "definitionId") @PathParam("definitionId") @NotBlank String definitionId,
                            EventDefinitionDto dto) {
+        checkPermission(RestPermissions.EVENT_DEFINITIONS_EDIT);
         dbService.get(definitionId)
                 .orElseThrow(() -> new NotFoundException("Event definition <" + definitionId + "> doesn't exist"));
 
@@ -141,6 +146,7 @@ public class EventDefinitionsResource extends RestResource implements PluginRest
     @ApiOperation("Delete event definition")
     @AuditEvent(type = EventsAuditEventTypes.EVENT_DEFINITION_DELETE)
     public void delete(@ApiParam(name = "definitionId") @PathParam("definitionId") @NotBlank String definitionId) {
+        checkPermission(RestPermissions.EVENT_DEFINITIONS_DELETE);
         eventDefinitionHandler.delete(definitionId);
     }
 
@@ -150,6 +156,7 @@ public class EventDefinitionsResource extends RestResource implements PluginRest
     @AuditEvent(type = EventsAuditEventTypes.EVENT_DEFINITION_EXECUTE)
     public void execute(@ApiParam(name = "definitionId") @PathParam("definitionId") @NotBlank String definitionId,
                         @ApiParam(name = "parameters", required = true) @NotNull EventProcessorParameters parameters) {
+        checkPermission(RestPermissions.EVENT_DEFINITIONS_EXECUTE);
         if (parameters instanceof EventProcessorParametersWithTimerange.FallbackParameters) {
             throw new BadRequestException("Unknown parameters type");
         }
@@ -167,6 +174,7 @@ public class EventDefinitionsResource extends RestResource implements PluginRest
     @ApiOperation(value = "Validate an event definition")
     public ValidationResult validate(@ApiParam(name = "JSON body", required = true)
                                      @Valid @NotNull EventDefinitionDto toValidate) {
+        checkPermission(RestPermissions.EVENT_DEFINITIONS_CREATE);
         return toValidate.config().validate();
     }
 }
